requirements.txt
chromadb
langchain
beautifulsoup4
atlassian-python-api
langchain-openai
sentence-transformers
langchain-huggingface
langchain_chroma